refactor: move const folder in page-block
