+++ packages/blocks/src/common/bracket-pairs.ts
