+++ packages/blocks/src/page-block/types.ts
@@ -1,4 +1,3 @@
-import type { AFFINE_DATABASE_CONVERT_WIDGET_TAG } from '../widgets/database-convert/database-convert.js';
 import type { AFFINE_FORMAT_BAR_WIDGET_TAG } from '../widgets/format-bar/format-bar.js';
 import type { AFFINE_REMOTE_SELECTION_WIDGET_TAG } from '../widgets/remote-selection/remote-selection.js';
 import type { DocPageBlockComponent } from './doc/doc-page-block.js';
@@ -10,14 +9,12 @@ export type DocPageBlockWidgetName =
   | 'draggingArea'
   | 'dragHandle'
   | typeof AFFINE_FORMAT_BAR_WIDGET_TAG
-  | typeof AFFINE_DATABASE_CONVERT_WIDGET_TAG
   | typeof AFFINE_REMOTE_SELECTION_WIDGET_TAG;
 export type EdgelessPageBlockWidgetName =
   | 'slashMenu'
   | 'linkedPage'
   | 'dragHandle'
   | typeof AFFINE_FORMAT_BAR_WIDGET_TAG
-  | typeof AFFINE_DATABASE_CONVERT_WIDGET_TAG
   | typeof AFFINE_REMOTE_SELECTION_WIDGET_TAG;
 
 export type PageBlockComponent =
