+++ packages/blocks/src/preset/index.ts
@@ -21,7 +21,6 @@ import {
 import { PageBlockSchema } from '../page-block/page-model.js';
 import { ParagraphBlockSchema } from '../paragraph-block/paragraph-model.js';
 import { SurfaceBlockSchema } from '../surface-block/surface-model.js';
-import { AFFINE_DATABASE_CONVERT_WIDGET_TAG } from '../widgets/database-convert/database-convert.js';
 import { AFFINE_FORMAT_BAR_WIDGET_TAG } from '../widgets/format-bar/format-bar.js';
 import { AFFINE_REMOTE_SELECTION_WIDGET_TAG } from '../widgets/remote-selection/remote-selection.js';
 
@@ -37,9 +36,6 @@ const pageBlockSpec: LitBlockSpec<DocPageBlockWidgetName> = {
       dragHandle: literal`affine-drag-handle-widget`,
       [AFFINE_FORMAT_BAR_WIDGET_TAG]: literal`${unsafeStatic(
         AFFINE_FORMAT_BAR_WIDGET_TAG
-      )}`,
-      [AFFINE_DATABASE_CONVERT_WIDGET_TAG]: literal`${unsafeStatic(
-        AFFINE_DATABASE_CONVERT_WIDGET_TAG
       )}`,
       [AFFINE_REMOTE_SELECTION_WIDGET_TAG]: literal`${unsafeStatic(
         AFFINE_REMOTE_SELECTION_WIDGET_TAG
@@ -59,9 +55,6 @@ const edgelessBlockSpec: LitBlockSpec<EdgelessPageBlockWidgetName> = {
       dragHandle: literal`affine-drag-handle-widget`,
       [AFFINE_FORMAT_BAR_WIDGET_TAG]: literal`${unsafeStatic(
         AFFINE_FORMAT_BAR_WIDGET_TAG
-      )}`,
-      [AFFINE_DATABASE_CONVERT_WIDGET_TAG]: literal`${unsafeStatic(
-        AFFINE_DATABASE_CONVERT_WIDGET_TAG
       )}`,
       [AFFINE_REMOTE_SELECTION_WIDGET_TAG]: literal`${unsafeStatic(
         AFFINE_REMOTE_SELECTION_WIDGET_TAG
