+++ packages/blocks/src/widgets/format-bar/components/inline-items.ts
@@ -4,8 +4,7 @@ import type { AffineTextAttributes } from '../../../__internal__/rich-text/virgo
 import {
   includeInlineSupportedBlockSelected,
   inlineFormatConfig,
-} from '../../../page-block/const/inline-format-config.js';
-import { isPageComponent } from '../../../page-block/utils/guard.js';
+} from '../../../common/inline-format-config.js';
 import {
   getCombinedFormatInBlockSelections,
   getCombinedFormatInTextSelection,
@@ -14,19 +13,16 @@ import type { AffineFormatBarWidget } from '../format-bar.js';
 import { BackgroundButton } from './background/background-button.js';
 
 export const InlineItems = (formatBar: AffineFormatBarWidget) => {
-  const pageElement = formatBar.pageElement;
-  if (!isPageComponent(pageElement)) {
-    throw new Error('the pageElement of formatBar is not a PageComponent');
-  }
+  const root = formatBar.root;
 
-  if (!includeInlineSupportedBlockSelected(pageElement)) {
+  if (!includeInlineSupportedBlockSelected(root)) {
     return nothing;
   }
 
   let type: 'text' | 'block' = 'text';
   let format: AffineTextAttributes = {};
-  const textSelection = pageElement.selection.find('text');
-  const blockSelections = pageElement.selection.filter('block');
+  const textSelection = root.selectionManager.find('text');
+  const blockSelections = root.selectionManager.filter('block');
 
   if (
     !(
@@ -38,20 +34,17 @@ export const InlineItems = (formatBar: AffineFormatBarWidget) => {
   }
 
   if (textSelection) {
-    format = getCombinedFormatInTextSelection(pageElement.root, textSelection);
+    format = getCombinedFormatInTextSelection(root, textSelection);
     type = 'text';
   } else {
-    format = getCombinedFormatInBlockSelections(
-      pageElement.root,
-      blockSelections
-    );
+    format = getCombinedFormatInBlockSelections(root, blockSelections);
     type = 'block';
   }
 
   const backgroundButton = BackgroundButton(formatBar);
 
   return html`${inlineFormatConfig
-      .filter(({ showWhen }) => showWhen(pageElement))
+      .filter(({ showWhen }) => showWhen(root))
       .map(
         ({ id, name, icon, action, activeWhen }) =>
           html`<icon-button
@@ -61,7 +54,7 @@ export const InlineItems = (formatBar: AffineFormatBarWidget) => {
             ?active=${activeWhen(format)}
             @click=${() => {
               action({
-                blockElement: pageElement,
+                root,
                 type,
                 format,
               });
